Add functions to find "representative reefs"
Functions to cluster intervention reefs according to distance into clusters which represent groups of reefs with a maximum distance apart. Reefs in different clusters accrue travel to port costs as it is assumed the clusters are too far apart for a ship not to travel back to port in between.